+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -6,13 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.9.0] - 2019-25-11
+## [0.9.2] - 2020-Apr-25
 
 ### Added
--
+
+- Properties to Tilesets.
+- Test verifying `tileset.properties`
 
 ### Changed
-- Migration to `rust 2018` and `?`
+
+-
 
 ### Removed
+
 -